+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.8.0
+
+- Extracted caching of Heroku worker counts and added experimental Redis cache
+    scaler.counter_cache = Autoscaler::CounterCacheRedis(Sidekiq.method(:redis))
+
 ## 0.7.0
 
 - Added Autoscaler::LinearScalingStrategy